Refactored code to follow architecture. Updated AuthService so it considers user.status. Created UserResponse schema. Created enums.py
