+++ app/schemas/enums.py
@@ -1,0 +1,6 @@
+"""App/Database enums"""
+from enum import Enum
+
+class Status(str, Enum):
+    ACTIVE = "ACTIVE"
+    INACTIVE = "INACTIVE"
