+++ app/services/AuthService.py
@@ -5,6 +5,7 @@ from app.services.utils.hash.Hash import Hash
 from app.services.utils.hash.HashCrypt import HashCrypt
 from app.database.database import SessionLocal
 from app.database.models import User
+from app.schemas.enums import Status
 
 
 
@@ -15,7 +16,7 @@ class AuthService:
     def login(self, username, password) -> bool:
         """Check if a user has the correct credentials to login"""
         with SessionLocal() as session:
-            user = session.query(User).filter(User.username == username).first()
+            user = session.query(User).filter(User.username == username, User.status == Status.ACTIVE).first()
             if not user:
                 return False
             if not self.hash_service.verify(password, user.password):
